PR updates
* Combined translation utilities into one script
* Move cql-translation-service-client to dependencies
* Add notes to ValueSetMapper
