+++ package.json
@@ -20,7 +20,6 @@
   "devDependencies": {
     "cql-exec-fhir": "^1.3.0",
     "cql-execution": "^1.4.1",
-    "cql-translation-service-client": "^0.6.0",
     "eslint": "^7.20.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
@@ -28,5 +27,8 @@
     "prettier": "^2.2.1"
   },
   "author": "",
-  "license": "Apache-2.0"
+  "license": "Apache-2.0",
+  "dependencies": {
+    "cql-translation-service-client": "^0.6.0"
+  }
 }
